// src/app/model/PostTransResponse.ts
export class PostTransResponse {  
    totalPostedSucess: number;
    totalPostedFailed: number;
    successMsg: string;
    errorMsg: string;
    transTracer: string;
    batchNo: number;
    postErrorCode: number;
    postErrorText: string;
    transTracerCount: number;
    transReference: string;
}
        